refactor: change tasks data structure
